refactor(react): move appearance provider + editor UI into ui/appearance
Consolidates all appearance code into packages/react/src/components/ui/appearance/
so the export script (#541) can copy ui/ wholesale without special-casing.

- provider/ — NexusAppearanceProvider, factory, script, server (from src/appearance/)
- appearance-settings/, color-field/, setting-row/, config-preview/, theme-quick-control/
  — editor UI moved from apps/console/src/modules/design-system/appearance/
- Editor UI components exported from the main @nexus/react entry
- Console imports updated to use @nexus/react directly (no owned copies)
- vite entry points and storybook preview updated to new provider path

Closes #539 item C.
