+++ packages/react/src/components/ui/appearance/appearance-settings/index.ts
@@ -1,0 +1,1 @@
+export { NexusAppearanceSettings } from './appearance-settings';
