+++ packages/react/src/components/ui/appearance/color-field/color-field.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 
-import { Input } from '@nexus/react';
+import { Input } from '@/components/ui/input';
 
 interface NexusAppearanceColorFieldProps {
   value: string;
