+++ packages/react/src/components/ui/appearance/color-field/index.ts
@@ -1,0 +1,1 @@
+export { NexusAppearanceColorField } from './color-field';
