+++ packages/react/src/components/ui/appearance/config-preview/index.ts
@@ -1,0 +1,2 @@
+export type { NexusAppearanceConfigPreviewProps } from './config-preview';
+export { NexusAppearanceConfigPreview } from './config-preview';
