+++ packages/react/src/components/ui/appearance/setting-row/index.ts
@@ -1,0 +1,1 @@
+export { NexusAppearanceSettingRow } from './setting-row';
