+++ packages/react/src/components/ui/appearance/theme-quick-control/index.ts
@@ -1,3 +1,2 @@
-export { NexusAppearanceSettings } from './appearance-settings';
 export type { NexusThemeQuickControlProps } from './theme-quick-control';
 export { NexusThemeQuickControl } from './theme-quick-control';
